API Reoganising the state directory
