+++ client/src/state/editor/elementalConfigProvider.js
@@ -1,0 +1,8 @@
+import Config from 'lib/Config';
+
+/**
+ * Returns the client config provided by `ElementalAreaController`
+ *
+ * @returns {object}
+ */
+export default () => Config.getSection('DNADesign\\Elemental\\Controllers\\ElementalAreaController');
